refactor: extract page sections into feature components
Break accounts, profile, and ramping pages into self-contained section
components under components/features/{Accounts,Profile,Routes}/. Each
section owns its own state, data fetching, and modals. Page files are
now thin orchestrators (~15-30 lines each).
