+++ components/features/Accounts/index.ts
@@ -1,0 +1,3 @@
+export { default as BankAccountsSection } from './BankAccountsSection';
+export { default as SafeWalletsSection } from './SafeWalletsSection';
+export type { BankAccount, SafeWallet } from './types';
