+++ components/features/Profile/index.ts
@@ -1,0 +1,3 @@
+export { default as ProfileFormSection } from './ProfileFormSection';
+export { default as LinkedWalletsSection } from './LinkedWalletsSection';
+export { default as ApiKeysSection } from './ApiKeysSection';
